#49695: add enableCross option, formatting, split component into separate files
