+++ components/03-modules/range/Range.stories.js
@@ -1,18 +1,18 @@
-import {storiesOf} from '@storybook/vue'
-import {action} from "@storybook/addon-actions"
+import { storiesOf } from '@storybook/vue'
+import { action } from "@storybook/addon-actions"
 
 import App from '../../01-globals/app/App.vue'
-import Range from './Range.vue'
+import AlpacaRange from './Range.vue'
 
 storiesOf('Modules/Range', module).add('Default', () => ({
-  components: { App, Range },
+  components: { App, AlpacaRange },
   template: `
     <app>
-      <range 
-      v-model="value" 
-      :min="min"
-      :max="max"
-      @apply="updateVal"
+      <alpaca-range
+        v-model="value" 
+        :min="min"
+        :max="max"
+        @apply="updateVal"
       />
       Range: {{ value }}
     </app>
